finish adding tests that all pass
